DFWannier Documentation

DFWannier Documentation

This package is meant to be used as a post processing tool with Wannier90 output. Mainly to calculate various model properties such as Heisenberg exchanges.

Exchanges

DFWannier.DHvecvals — Function.

DHvecvals(hami::TbHami{T}, k_grid, atoms)

Calculates $D(k) = [H(k), J]$, $P(k)$ and $L(k)$ where $H(k) = P(k) L(k) P^{-1}(k)$. hami should be a tuple containing the $H_{up}, H_{down}$ blocks of the full Hamiltonian.

source

DHvecvals(hami::TbHami{T}, k_grid, atoms)

Calculates $D(k) = [H(k), J]$, $P(k)$ and $L(k)$ where $H(k) = P(k) L(k) P^{-1}(k)$. hami should be the full Hamiltonian containing both spin-diagonal and off-diagonal blocks.

source